requirements.txt
# Core ML libraries
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0
peft>=0.10.0
trl>=0.8.0

# Data handling
datasets>=2.18.0
pandas>=2.2.0
numpy>=1.26.0

# Tokenizers
sentencepiece>=0.2.0
protobuf>=4.25.0

# Training utilities
tensorboard>=2.16.0
wandb>=0.16.0
scipy>=1.12.0

# Configuration
pyyaml>=6.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0